Match on array instead of cons
In Scala 2.13 a new type is introduced, the arrayseq https://www.scala-lang.org/api/2.13.1/scala/collection/immutable/ArraySeq.html
At runtime, the return type of schema.metricbuilder.name is an ArraySeq in 2.13.

The match for the default pattern, uses the cons type (`::`), which is not applicable to ArraySeq,
and caused a MatchError of the partial function.

By changing the match to a list deconstructor (`+:`) this is circumvented, and it also works in 2.12.

Tests have been added to show this behaviour